+++ jax/experimental/jax2tf/jax2tf.py
@@ -304,6 +304,7 @@ def convert(fun_jax: Callable,
       so the lowering tries harder to use non-XLA TF ops to lower the
       function and aborts if this is not possible. Cannot be set to `False`
       when using `native_serialization`.
+      Starting with JAX 0.4.31 support for `enable_xla=False` is deprecated.
     native_serialization: serialize the JAX function natively to
       StableHLO with compatibility guarantees. This makes it easier to have
       confidence that the code executed when calling this function from
@@ -312,6 +313,7 @@ def convert(fun_jax: Callable,
       is set to `False` or to the configuration flag
       `--jax2tf_default_native_serialization` otherwise.
       Native serialization cannot be used with `enable_xla=False`.
+      Starting with JAX 0.4.31 support for non-native serialization is deprecated.
     native_serialization_platforms: In conjunction with
       `native_serialization`, specify the platform(s)
       for which to lower the code. Must be a tuple of
@@ -327,12 +329,17 @@ def convert(fun_jax: Callable,
     tuple/lists/dicts thereof), and returns TfVals as outputs, and uses
     only TensorFlow ops and thus can be called from a TensorFlow program.
   """
+  if not enable_xla:
+    warnings.warn("jax2tf.convert with enable_xla=False is deprecated.")
   if native_serialization is DEFAULT_NATIVE_SERIALIZATION:
     if not enable_xla:
       native_serialization = False
     else:
       native_serialization = config.jax2tf_default_native_serialization.value
 
+  if not native_serialization:
+    warnings.warn(
+        "jax2tf.convert with native_serialization=False is deprecated.")
   if native_serialization and not enable_xla:
     raise ValueError(
         "native_serialization is not supported with enable_xla=False")